requirements.txt
cyvcf2
psycopg2-binary
elasticsearch
confluent_kafka
